refactored waterfall state to context. Created ListContext component in context folder. Removed passed propos from components
